reorganize local vs general utils
